Add dapr run  flag to make the app exits gracefully

+++ cmd/run.go
@@ -20,6 +20,7 @@ import (
 	"runtime"
 	"strconv"
 	"strings"
+	"syscall"
 	"time"
 
 	"github.com/spf13/cobra"
@@ -47,6 +48,7 @@ var (
 	metricsPort        int
 	maxRequestBodySize int
 	unixDomainSocket   string
+	appGraceExit       bool
 )
 
 const (
@@ -108,6 +110,7 @@ var RunCmd = &cobra.Command{
 			MetricsPort:        metricsPort,
 			MaxRequestBodySize: maxRequestBodySize,
 			UnixDomainSocket:   unixDomainSocket,
+			AppGraceExit:       appGraceExit,
 		})
 		if err != nil {
 			print.FailureStatusEvent(os.Stderr, err.Error())
@@ -306,7 +309,12 @@ var RunCmd = &cobra.Command{
 		}
 
 		if output.AppCMD != nil && (output.AppCMD.ProcessState == nil || !output.AppCMD.ProcessState.Exited()) {
-			err = output.AppCMD.Process.Kill()
+			if output.AppGraceExit {
+				err = output.AppCMD.Process.Signal(syscall.SIGTERM)
+			} else {
+				err = output.AppCMD.Process.Kill()
+			}
+
 			if err != nil {
 				print.FailureStatusEvent(os.Stderr, fmt.Sprintf("Error exiting App: %s", err))
 			} else {
@@ -340,6 +348,7 @@ func init() {
 	RunCmd.Flags().BoolP("help", "h", false, "Print this help message")
 	RunCmd.Flags().IntVarP(&maxRequestBodySize, "dapr-http-max-request-size", "", -1, "Max size of request body in MB")
 	RunCmd.Flags().StringVarP(&unixDomainSocket, "unix-domain-socket", "u", "", "Path to a unix domain socket dir. If specified, Dapr API servers will use Unix Domain Sockets")
+	RunCmd.Flags().BoolVar(&appGraceExit, "app-grace-exit", false, "Enable graceful exit for the application")
 
 	RootCmd.AddCommand(RunCmd)
 }
